uefi-sct/SctPkg: allowable NotifyTpl in CreateEvent
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=3057

According to table Table 27. TPL Restrictions in UEFI spec 2.8 B:

Event Notification Levels must be "> TPL_APPLICATION" and "<= TPL_HIGH_LEVEL".

In the conformance test for CreateEvent() check that TPL_APPLICATION
results in EFI_INVALID_PARAMETER if the event can be notified.

In the functional tests check that events can be created with
TPL_HIGH_LEVEL.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>